% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Oncology trial cutoff data}
\format{
This data frame contains the following columns:
\describe{
\item{ID}{patient ID assigned by the trial conductor}
\item{evl_time.t0}{time of the baseline visit in month}
\item{evl_time.t1}{time of the 1st visit in month}
\item{evl_time.t2}{time of the 2nd visit in month}
\item{evl_time.t3}{time of the 3rd visit in month}
\item{evl_time.t4}{time of the 4th visit in month}
\item{evl_time.t5}{time of the 5th visit in month}
\item{evl_time.t6}{time of the 6th visit in month}
\item{evl_time.t7}{time of the 7th visit in month}
\item{evl_time.t8}{time of the 8th visit in month}
\item{evl_time.t9}{time of the 9th visit in month}
\item{evl_time.t10}{time of the 10th visit in month}
\item{evl_time.t11}{time of the 11th visit in month}
\item{evl_time.t12}{time of the 12th visit in month}
\item{evl_time.t13}{time of the 13th visit in month}
\item{evl_time.t14}{time of the 14th visit in month}
\item{evl_time.t15}{time of the 15th visit in month}
\item{mu.t0}{measurements of the target lesion at the baseline visit}
\item{mu.t1}{measurements of the target lesion at the 1st visit}
\item{mu.t2}{measurements of the target lesion at the 2nd visit}
\item{mu.t3}{measurements of the target lesion at the 3rd visit}
\item{mu.t4}{measurements of the target lesion at the 4th visit}
\item{mu.t5}{measurements of the target lesion at the 5th visit}
\item{mu.t6}{measurements of the target lesion at the 6th visit}
\item{mu.t7}{measurements of the target lesion at the 7th visit}
\item{mu.t8}{measurements of the target lesion at the 8th visit}
\item{mu.t9}{measurements of the target lesion at the 9th visit}
\item{mu.t10}{measurements of the target lesion at the 10th visit}
\item{mu.t11}{measurements of the target lesion at the 11th visit}
\item{mu.t12}{measurements of the target lesion at the 12th visit}
\item{mu.t13}{measurements of the target lesion at the 13th visit}
\item{mu.t14}{measurements of the target lesion at the 14th visit}
\item{mu.t15}{measurements of the target lesion at the 15th visit}
\item{time_PFS}{time to progression-free survival progressive disease}
\item{delta_PFS}{censored status of \code{time_PFS}, 1 - observed, 2 - censored}
\item{time_NL}{time to new lesion progressive disease}
\item{delta_NL}{censored status of \code{time_NL}, 1 - observed, 2 - censored}
\item{time_NT}{time to non-target lesion progressive disease}
\item{delta_NT}{censored status of \code{time_NT}, 1 - observed, 2 - censored}
\item{time_OS}{time to death}
\item{delta_OS}{censored status of \code{time_OS}, 1 - observed, 2 - censored}
\item{trt}{assigned treatment}
\item{rand_date}{participant's randomization date after enrolled in the trial}
}
}
\usage{
data
}
\description{
sample dataset of an oncology trial with cutoff at any time point. In total
400 patients are included.
}
\examples{
\donttest{
 result = osbma(osbma::data, covariate = "trt")
}

}
\keyword{data}
