% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osbma.R
\name{osbma}
\alias{osbma}
\title{Prediction of overall survival through Bayesian model averaging}
\usage{
osbma(
  data,
  covariate,
  longest.survival = 365 * 20/30,
  n.MCMC.chain = 2,
  no.adapt = 80,
  burn.in = 10,
  MCMC.sample = 100,
  thin = 10,
  method = "rjparallel",
  beta.prior.sd = 10,
  alpha.prior = 1,
  eta.prior = 0.5,
  lambda.prior.sd = 10,
  b.prior.sd = 1,
  y.prior.sd = 1,
  ...
)
}
\arguments{
\item{...}{}
}
\value{
\item{}{}
}
\description{
This function improves the accuracy of the OS forecast by combining joint
models developed based on each granular component of PFS through Bayesian
model averaging.
}
