% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osbma.R
\name{plot.predict.osbma}
\alias{plot.predict.osbma}
\title{plot based on prediction outcomes of an osbma object}
\usage{
\method{plot}{predict.osbma}(object, trt.col.name = "trt", type = "date", ...)
}
\arguments{
\item{...}{}
}
\description{
plot based on prediction outcomes of an osbma object
}
