% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osbma.R
\name{predict.osbma}
\alias{predict.osbma}
\title{overall survival prediction}
\usage{
\method{predict}{osbma}(object, quantile, ...)
}
\arguments{
\item{...}{}
}
\description{
\code{predict} method for class "\code{osbma}"
}
