% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osbma.R
\name{osbma}
\alias{osbma}
\alias{summary.osbma}
\alias{print.osbma}
\alias{predict.osbma}
\alias{plot.predict.osbma}
\title{Prediction of overall survival through Bayesian model averaging}
\usage{
osbma(
  data,
  covariate,
  longest_survival = 2000,
  n.MCMC.chain = 2,
  no.adapt = 20000,
  burn.in = 10000,
  MCMC.sample = 30000,
  thin = 3,
  method = "rjparallel",
  beta.prior.sd = 1,
  alpha.prior = 1,
  eta.prior = 1,
  lambda.prior.sd = 1,
  b.prior.sd = 1,
  y.prior.sd = 0.1,
  ...
)

\method{summary}{osbma}(object, ...)

\method{print}{osbma}(x, ...)

\method{predict}{osbma}(object, quantile = 0.9, ...)

\method{plot}{predict.osbma}(x, trt.col.name = "trt", type = "date", ...)
}
\arguments{
\item{data}{trial dataset with progression-free survival info (measurements
of target lesions, time to non-target lesion, time to new lesion, time to
death), covaraites, patient ID, randomization date, etc. Please see
\code{\link{data}} for more detailed descriptions.}

\item{covariate}{a vector of column names of all the covariates.}

\item{longest_survival}{the longest possible survival of a patient in this
trial.}

\item{n.MCMC.chain}{number of MCMC chains, default to 2.}

\item{no.adapt}{the number of iterations for adaptation.}

\item{burn.in}{number of burin-in iterations for MCMC.}

\item{MCMC.sample}{number of iterations for MCMC.}

\item{thin}{thinning interval for monitors.}

\item{method}{same as the \code{method} argument in \code{\link[runjags]{run.jags}},
the method with which to call JAGS; probably a character vector specifying
one of 'rjags', 'simple', 'interruptible', 'parallel', 'rjparallel',
'background', 'bgparallel' or 'snow'. The 'rjags' and 'rjparallel' methods
run JAGS using the rjags package, whereas other options do not require the
rjags package and call JAGS as an external executable.}

\item{beta.prior.sd}{the standard deviation of the normal priors of all betas
(coefficient of covariates)}

\item{alpha.prior}{the rate of the exponential priors of all alphas (the shape
parameter)}

\item{eta.prior}{the rate of the exponential priors of all etas (correlations
in the Clayton copula)}

\item{lambda.prior.sd}{the standard deviation of the normal priors of all
lambdas}

\item{b.prior.sd}{the standard deviation of the normal priors of all random
effects \code{b}}

\item{y.prior.sd}{the standard deviation of the normal prior of Y (the observed
lesion measurement)}

\item{...}{further arguments. Not currently used.}

\item{object}{an object of class "\code{osbma}", usually, a result of a call to
\code{\link{osbma}}}

\item{x}{an object of class "\code{predict.osbma}", usually, a result of a call to
\code{\link{predict.osbma}}}

\item{quantile}{the quantile of the predicted overall survival}

\item{trt.col.name}{the name of the assigned treatment column, default to \code{trt}}

\item{type}{type of plot the users want: "date" generates a plot with patients'
survival timeline; "KM" generates a Kaplan-Meier curve of overall survival;
a number n between 0 - N, where N is the total number of patients enrolled,
generates a plot of the estimated time of the nth death in the trial.}
}
\value{
\item{posterior_sample}{an \code{runjags} object generated through the \code{runjags::run.jags}
function, which includes posterior samples of all the parameters and predicted
survival times}
\item{rand_date}{the randomization dates of patients}
\item{covariate}{the covariates columns of the input dataset}
\item{OS}{the overall survival and its censored status provided in the input
dataset}
\item{ID}{the patient ID}

The summary method returns a numeric matrix of summary statistics for each
variable. Please check \code{\link[runjags]{summary.runjags}} for more
details.

\item{prediction}{the predicted overall survival based on \code{osbma} fit}
\item{prediction_date}{the death date predicted by the \code{osbma} model}
\item{quantile}{the quantile specified by the user}
\item{covariate}{the columns of covariates provided in the input dataset}
\item{rand_date}{the randomization date of the patients in the trial}
\item{OS}{the overall survival and its censored status provided in the input
dataset}
\item{ID}{the ID of the patients provided in the input dataset}

3 types of plots
}
\description{
This function improves the accuracy of the OS forecast by combining joint
models developed based on each granular component of PFS through Bayesian
model averaging.

\code{summary} method for class "\code{osbma}"

\code{predict} method for class "\code{osbma}"
}
\examples{
\donttest{
result <- osbma(osbma::data, covariate = "trt")
}

}
\references{
(manuscript under review)
}
\seealso{
\code{\link{predict.osbma}} \cr
\code{\link{plot.predict.osbma}} \cr
\code{\link[runjags]{run.jags}}
}
